BigTenX Admin Panel - Complete Platform
Features:
- Full admin dashboard with light mode
- User management with search and pagination
- Vault management with 5 tabs
- Task management with create/edit/delete and analytics
- Marketplace with product CRUD
- Notifications with send/schedule
- Contests with leaderboard
- XP & Levels configuration
- Referral system with leaderboard
- Badges & VIP tiers
- Carousel manager
- Analytics dashboard
- Mobile responsive layout
- User profile with earnings
- Wallet with deposit/withdraw/history
- Premium fin-tech grade UI
- Dark mode support
- API integration with PHP backend

+++ api/tasks/hot_offers.php
@@ -27,6 +27,7 @@ try {
         ? "AND (t.max_users IS NULL OR {$completionsExpr} < t.max_users)"
         : "";
 
+    // ─── FIX: Include expires_at in SELECT ────────────────────────────────
     $sql = "
         SELECT
             t.id,
@@ -54,11 +55,12 @@ try {
     $tasks = $db->query($sql)->fetchAll(PDO::FETCH_ASSOC);
 
     foreach ($tasks as &$t) {
-        $t['reward_xp']        = (int)$t['reward_xp'];
+        $t['reward_xp']        = (float)$t['reward_xp'];  // Changed to float for decimal support
         $t['apply_multiplier'] = (int)$t['apply_multiplier'];
         $t['completions']      = (int)$t['completions'];
         $t['max_users']        = $t['max_users'] !== null ? (int)$t['max_users'] : null;
 
+        // ─── FIX: Decode steps properly ──────────────────────────────────
         $t['steps'] = (!empty($t['steps']))
             ? (is_array($decoded = json_decode($t['steps'], true)) ? $decoded : [])
             : [];
